+++ firedecomp/benders/utils.py
@@ -106,13 +106,13 @@ def get_start_resource_info(data, i, start_period):
     rest = {(i, t): 0 for t in data.T}
     travel = {(i, t): 0 for t in data.T}
 
-    if (data.ITW[i] is not True) and (data.IOW[i] is not True):
-        work_count = 1
-    else:
+    if (data.ITW[i] == True) or (data.IOW[i] == True):
         if start_period == data.min_t:
             work_count = data.CWP[i] - data.CRP[i] + 1
         else:
             work_count = data.WP[i] + 1
+    else:
+        work_count = 1
 
     start = False
     travel_count = 1
@@ -121,7 +121,7 @@ def get_start_resource_info(data, i, start_period):
             start = True
 
         if start:
-            if t == data.max_t:
+            if t >= data.max_t - data.TRP[i] + 1:
                 work_count += 1
                 travel_count += 1
                 travel[i, t] = 1
@@ -142,18 +142,15 @@ def get_start_resource_info(data, i, start_period):
                     work[i, t] = 1
             elif work_count <= data.WP[i]:
                 work_count += 1
-                travel_count += 1
                 travel[i, t] = 1
             elif work_count <= data.WP[i] + data.RP[i]:
                 work_count += 1
                 rest[i, t] = 1
-            elif work_count < data.WP[i] + data.RP[i]:
+            elif work_count < data.WP[i] + data.RP[i] + data.TRP[i]:
                 work_count += 1
-                travel_count += 1
                 travel[i, t] = 1
             else:
-                work_count = 0
-                travel_count += 1
+                work_count = 1 + data.TRP[i]
                 travel[i, t] = 1
         else:
             pass