+++ simulations/execute.sh
@@ -1,0 +1,1 @@
+firedecomp_simulations -n 25 -nb 5 10 20 -na 5 10 20 -nm 5 10 20 -np 20 30 40 -m original benders -so solver_options.yaml -vvv &> simulations.log