+++ ChangeLog.md
@@ -19,6 +19,38 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed ###
 
 
+1.1.0 - 2019-02-22
+------------------
+
+### Added ###
+- Add options to solve the original feasible model adding slack variables to wildfire containment constraints.
+- Add simulations folder.
+
+### Changed ###
+- Round to 3 digits model information to avoid numerical problems.
+
+### Removed ###
+
+### Fixed ###
+- Fix error computing rests periods.
+
+
+1.0.0 - 2019-02-18
+------------------
+
+### Added ###
++ Add reformulation
++ Add command line option to execute simulations.
+
+### Changed ###
++ Update classes structure.
+
+### Removed ###
+
+### Fixed ###
++ Fix model errors.
+
+
 0.1.0 - 2019-01-14
 ------------------